test(cypress): fix move test support file commands-components.js
